Upgrade build to rebar3; Upgrade deps
We've changed the build tool from rebar to rebar3. We've also upgraded
the jsx dependency and changed the protobuf dependency from Basho's
erlang_protobuff, which is no longer maintained, to gpb which is
included with the rebar3_gpb_plugin. Changing the protobuf library
required some changes to the record format but no semantic changes
were required.